requirements.txt
nest_asyncio==1.6.0
numpy==1.26.2
pandas==2.3.0
pydantic==2.11.7
python-dotenv==1.1.1
tqdm==4.67.1

# Vertex AI / Gemini
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# RAG dependencies (local HuggingFace embeddings, no API key needed)
langchain>=0.3.0
langchain_community>=0.3.0
langchain_huggingface>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4  # use faiss-gpu if you have CUDA

# Optional: for open-source models via vLLM
# vllm==0.8.3
# transformers==4.52.4
